backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pgvector
python-dotenv
python-multipart
Pillow
torch
transformers
requests